+++ internal/output/group.go
@@ -5,24 +5,24 @@ import (
 	"io"
 )
 
-type Group struct{
+type Group struct {
 	Begin, End string
 }
 
-func (g Group) WrapWriter(w io.Writer, _ string, tmpl Templater) io.Writer {
-	gw := &groupWriter{writer: w}
+func (g Group) WrapWriter(stdOut, _ io.Writer, _ string, tmpl Templater) (io.Writer, io.Writer, CloseFunc) {
+	gw := &groupWriter{writer: stdOut}
 	if g.Begin != "" {
 		gw.begin = tmpl.Replace(g.Begin) + "\n"
 	}
 	if g.End != "" {
 		gw.end = tmpl.Replace(g.End) + "\n"
 	}
-	return gw
+	return gw, gw, func() error { return gw.close() }
 }
 
 type groupWriter struct {
-	writer io.Writer
-	buff   bytes.Buffer
+	writer     io.Writer
+	buff       bytes.Buffer
 	begin, end string
 }
 
@@ -30,7 +30,7 @@ func (gw *groupWriter) Write(p []byte) (int, error) {
 	return gw.buff.Write(p)
 }
 
-func (gw *groupWriter) Close() error {
+func (gw *groupWriter) close() error {
 	if gw.buff.Len() == 0 {
 		// don't print begin/end messages if there's no buffered entries
 		return nil