+++ internal/output/output_test.go
@@ -16,7 +16,7 @@ import (
 func TestInterleaved(t *testing.T) {
 	var b bytes.Buffer
 	var o output.Output = output.Interleaved{}
-	var w = o.WrapWriter(&b, "", nil)
+	var w, _, _ = o.WrapWriter(&b, io.Discard, "", nil)
 
 	fmt.Fprintln(w, "foo\nbar")
 	assert.Equal(t, "foo\nbar\n", b.String())
@@ -27,14 +27,19 @@ func TestInterleaved(t *testing.T) {
 func TestGroup(t *testing.T) {
 	var b bytes.Buffer
 	var o output.Output = output.Group{}
-	var w = o.WrapWriter(&b, "", nil).(io.WriteCloser)
+	var stdOut, stdErr, cleanup = o.WrapWriter(&b, io.Discard, "", nil)
 
-	fmt.Fprintln(w, "foo\nbar")
+	fmt.Fprintln(stdOut, "out\nout")
 	assert.Equal(t, "", b.String())
-	fmt.Fprintln(w, "baz")
+	fmt.Fprintln(stdErr, "err\nerr")
 	assert.Equal(t, "", b.String())
-	assert.NoError(t, w.Close())
-	assert.Equal(t, "foo\nbar\nbaz\n", b.String())
+	fmt.Fprintln(stdOut, "out")
+	assert.Equal(t, "", b.String())
+	fmt.Fprintln(stdErr, "err")
+	assert.Equal(t, "", b.String())
+
+	assert.NoError(t, cleanup())
+	assert.Equal(t, "out\nout\nerr\nerr\nout\nerr\n", b.String())
 }
 
 func TestGroupWithBeginEnd(t *testing.T) {
@@ -53,19 +58,19 @@ func TestGroupWithBeginEnd(t *testing.T) {
 	}
 	t.Run("simple", func(t *testing.T) {
 		var b bytes.Buffer
-		var w = o.WrapWriter(&b, "", &tmpl).(io.WriteCloser)
+		var w, _, cleanup = o.WrapWriter(&b, io.Discard, "", &tmpl)
 
 		fmt.Fprintln(w, "foo\nbar")
 		assert.Equal(t, "", b.String())
 		fmt.Fprintln(w, "baz")
 		assert.Equal(t, "", b.String())
-		assert.NoError(t, w.Close())
+		assert.NoError(t, cleanup())
 		assert.Equal(t, "::group::example-value\nfoo\nbar\nbaz\n::endgroup::\n", b.String())
 	})
 	t.Run("no output", func(t *testing.T) {
 		var b bytes.Buffer
-		var w = o.WrapWriter(&b, "", &tmpl).(io.WriteCloser)
-		assert.NoError(t, w.Close())
+		var _, _, cleanup = o.WrapWriter(&b, io.Discard, "", &tmpl)
+		assert.NoError(t, cleanup())
 		assert.Equal(t, "", b.String())
 	})
 }
@@ -73,7 +78,7 @@ func TestGroupWithBeginEnd(t *testing.T) {
 func TestPrefixed(t *testing.T) {
 	var b bytes.Buffer
 	var o output.Output = output.Prefixed{}
-	var w = o.WrapWriter(&b, "prefix", nil).(io.WriteCloser)
+	var w, _, cleanup = o.WrapWriter(&b, io.Discard, "prefix", nil)
 
 	t.Run("simple use cases", func(t *testing.T) {
 		b.Reset()
@@ -82,6 +87,7 @@ func TestPrefixed(t *testing.T) {
 		assert.Equal(t, "[prefix] foo\n[prefix] bar\n", b.String())
 		fmt.Fprintln(w, "baz")
 		assert.Equal(t, "[prefix] foo\n[prefix] bar\n[prefix] baz\n", b.String())
+		assert.NoError(t, cleanup())
 	})
 
 	t.Run("multiple writes for a single line", func(t *testing.T) {
@@ -92,7 +98,7 @@ func TestPrefixed(t *testing.T) {
 			assert.Equal(t, "", b.String())
 		}
 
-		assert.NoError(t, w.Close())
+		assert.NoError(t, cleanup())
 		assert.Equal(t, "[prefix] Test!\n", b.String())
 	})
 }